+++ src-docs/src/views/combo_box/containers.js
@@ -123,7 +123,9 @@ export default () => {
     modal = (
       <EuiModal onClose={closeModal} style={{ width: '800px' }}>
         <EuiModalHeader>
-          <EuiModalHeaderTitle>Combo box in a modal</EuiModalHeaderTitle>
+          <EuiModalHeaderTitle>
+            <h1>Combo box in a modal</h1>
+          </EuiModalHeaderTitle>
         </EuiModalHeader>
 
         <EuiModalBody>{comboBox}</EuiModalBody>