+++ src-docs/src/views/header/header_example.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import { renderToHtml } from '../../services';
-
 import { GuideSectionTypes } from '../../components';
 import lightColors from '!!sass-vars-to-js-loader!../../../../src/global_styling/variables/_colors.scss';
 
@@ -23,39 +21,30 @@ import { EuiHeaderSectionsProp } from './props';
 
 import Header from './header';
 const headerSource = require('!!raw-loader!./header');
-const headerHtml = renderToHtml(Header);
 
 import HeaderSections from './header_sections';
 const headerSectionsSource = require('!!raw-loader!./header_sections');
-const headerSectionsHtml = renderToHtml(HeaderSections);
 
 import HeaderPosition from './header_position';
 const headerPositionSource = require('!!raw-loader!./header_position');
-const headerPositionHtml = renderToHtml(HeaderPosition);
 
 import HeaderAlert from './header_alert';
 const headerAlertSource = require('!!raw-loader!./header_alert');
-const headerAlertHtml = renderToHtml(HeaderAlert);
 
 import HeaderAnimate from './header_animate';
 const headerAnimateSource = require('!!raw-loader!./header_animate');
-const headerAnimateHtml = renderToHtml(HeaderAnimate);
 
 import HeaderLinks from './header_links';
 const headerLinksSource = require('!!raw-loader!./header_links');
-const headerLinksHtml = renderToHtml(HeaderLinks);
 
 import HeaderDark from './header_dark';
 const headerDarkSource = require('!!raw-loader!./header_dark');
-const headerDarkHtml = renderToHtml(HeaderDark);
 
 import HeaderStacked from './header_stacked';
 const headerStackedSource = require('!!raw-loader!./header_stacked');
-const headerStackedHtml = renderToHtml(HeaderStacked);
 
 import HeaderElasticPattern from './header_elastic_pattern';
 const headerElasticPatternSource = require('!!raw-loader!./header_elastic_pattern');
-const headerElasticPatternHtml = renderToHtml(HeaderElasticPattern);
 
 const headerSnippet = `<EuiHeader>
   <EuiHeaderSection grow={false}>
@@ -167,23 +156,54 @@ export const HeaderExample = {
           type: GuideSectionTypes.JS,
           code: headerSource,
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerHtml,
-        },
       ],
       text: (
-        <p>
-          The header is made up of <strong>many</strong> individual components.
-        </p>
+        <>
+          <p>
+            The header is made up of <strong>many</strong> individual components
+            starting with <strong>EuiHeader</strong> as the container. You can
+            manually configure your header with the following related
+            components:
+          </p>
+          <ul>
+            <li>
+              <strong>EuiHeaderSection</strong>: Left/right containers with flex
+              properties.
+            </li>
+            <li>
+              <strong>EuiHeaderSectionItem</strong>: Containers for individual
+              header items as flex items.
+            </li>
+            <li>
+              <strong>EuiHeaderSectionItemButton</strong>: Specialized button
+              that extends{' '}
+              <Link to="/navigation/button#empty-button">
+                <strong>EuiButtonEmpty</strong>
+              </Link>{' '}
+              but styled to fit the height of the header with additional{' '}
+              <EuiCode>notification</EuiCode> props.
+            </li>
+            <li>
+              <strong>EuiHeaderLogo</strong>: A helpful component for creating a
+              linked logo that fits within the header sizing.
+            </li>
+            <li>
+              <strong>EuiHeaderBreadcrumbs</strong>: A set of{' '}
+              <Link to="/navigation/breadcrumbs">
+                <strong>EuiBreadcrumbs</strong>
+              </Link>{' '}
+              specifically stylized to fit inside the header.
+            </li>
+          </ul>
+        </>
       ),
       props: {
         EuiHeader,
-        EuiHeaderBreadcrumbs,
         EuiHeaderSection,
         EuiHeaderSectionItem,
         EuiHeaderSectionItemButton,
         EuiHeaderLogo,
+        EuiHeaderBreadcrumbs,
         EuiHeaderSectionsProp,
       },
       snippet: headerSnippet,
@@ -199,16 +219,12 @@ export const HeaderExample = {
           type: GuideSectionTypes.JS,
           code: headerSectionsSource,
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerSectionsHtml,
-        },
       ],
       text: (
         <>
           <p>
-            Alternatively, you can pass an array objects to the{' '}
-            <EuiCode>sections</EuiCode> props that takes a key of{' '}
+            Alternatively, you can pass an array of objects to the{' '}
+            <EuiCode>sections</EuiCode> prop that takes a key of{' '}
             <EuiCode>items</EuiCode> (array of children to wrap in an{' '}
             <strong>EuiHeaderSectionItem</strong>) and/or{' '}
             <EuiCode>breadcrumbs</EuiCode> (array of{' '}
@@ -232,6 +248,46 @@ export const HeaderExample = {
       },
       snippet: headerSectionsSnippet,
       demo: <HeaderSections />,
+      demoPanelProps: {
+        color: 'subdued',
+      },
+    },
+    {
+      title: 'Header links',
+      source: [
+        {
+          type: GuideSectionTypes.JS,
+          code: headerLinksSource,
+        },
+      ],
+      text: (
+        <>
+          <p>
+            In addition to the components mentioned prior, which lend themselves
+            to more application style navigation patterns,{' '}
+            <strong>EuiHeaderLinks</strong> and <strong>EuiHeaderLink</strong>{' '}
+            supply the ability to inline a list of navigational or menu style
+            links.
+          </p>
+          <p>
+            <strong>EuiHeaderLinks</strong> comes with responsive functionality
+            built-in which will convert the inline list of links to a popover
+            list triggered by a <strong>EuiHeaderSectionItemButton</strong>. You
+            can adjust at which breakpoints to switch to the popover display by
+            passing your own array of named breakpoints to{' '}
+            <EuiCode>popoverBreakpoints</EuiCode>.
+          </p>
+        </>
+      ),
+      props: {
+        EuiHeaderLinks,
+        EuiHeaderLink,
+      },
+      snippet: headerLinksSnippet,
+      demo: <HeaderLinks />,
+      demoPanelProps: {
+        color: 'subdued',
+      },
     },
     {
       title: 'Fixed header',
@@ -239,10 +295,6 @@ export const HeaderExample = {
         {
           type: GuideSectionTypes.JS,
           code: headerPositionSource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerPositionHtml,
         },
       ],
       text: (
@@ -269,42 +321,9 @@ export const HeaderExample = {
         '@include euiHeaderAffordForFixed;',
       ],
       demo: <HeaderPosition />,
-    },
-    {
-      title: 'Header links',
-      source: [
-        {
-          type: GuideSectionTypes.JS,
-          code: headerLinksSource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerLinksHtml,
-        },
-      ],
-      text: (
-        <>
-          <p>
-            <strong>EuiHeaderLinks</strong> and <strong>EuiHeaderLink</strong>{' '}
-            supply the ability to inline a list of navigational or menu style
-            links.
-          </p>
-          <p>
-            <strong>EuiHeaderLinks</strong> comes with responsive functionality
-            built-in which will convert the inline list of links to a popover
-            list triggered by a <strong>EuiHeaderSectionItemButton</strong>. You
-            can adjust at which breakpoints to switch to the popover display by
-            passing your own array of named breakpoints to{' '}
-            <EuiCode>popoverBreakpoints</EuiCode>.
-          </p>
-        </>
-      ),
-      props: {
-        EuiHeaderLinks,
-        EuiHeaderLink,
+      demoPanelProps: {
+        color: 'subdued',
       },
-      snippet: headerLinksSnippet,
-      demo: <HeaderLinks />,
     },
     {
       title: 'Dark theme',
@@ -312,10 +331,6 @@ export const HeaderExample = {
         {
           type: GuideSectionTypes.JS,
           code: headerDarkSource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerDarkHtml,
         },
       ],
       text: (
@@ -333,6 +348,9 @@ export const HeaderExample = {
       ),
       snippet: '<EuiHeader theme="dark" />',
       demo: <HeaderDark theme={lightColors} />,
+      demoPanelProps: {
+        color: 'subdued',
+      },
     },
     {
       title: 'Portal content in the header',
@@ -340,10 +358,6 @@ export const HeaderExample = {
         {
           type: GuideSectionTypes.JS,
           code: headerAlertSource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerAlertHtml,
         },
       ],
       text: (
@@ -388,6 +402,9 @@ export const HeaderExample = {
       },
       snippet: headerAlertSnippet,
       demo: <HeaderAlert />,
+      demoPanelProps: {
+        color: 'subdued',
+      },
     },
     {
       title: 'Header notifications',
@@ -395,10 +412,6 @@ export const HeaderExample = {
         {
           type: GuideSectionTypes.JS,
           code: headerAnimateSource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerAnimateHtml,
         },
       ],
       text: (
@@ -421,6 +434,9 @@ export const HeaderExample = {
       },
       snippet: headerAnimateSnippet,
       demo: <HeaderAnimate />,
+      demoPanelProps: {
+        color: 'subdued',
+      },
     },
     {
       title: 'Stacked headers',
@@ -428,10 +444,6 @@ export const HeaderExample = {
         {
           type: GuideSectionTypes.JS,
           code: headerStackedSource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerStackedHtml,
         },
       ],
       text: (
@@ -450,6 +462,9 @@ export const HeaderExample = {
         '@include euiHeaderAffordForFixed($euiHeaderHeightCompensation * 2);',
       ],
       demo: <HeaderStacked />,
+      demoPanelProps: {
+        color: 'subdued',
+      },
     },
     {
       title: 'The Elastic navigation pattern',
@@ -457,10 +472,6 @@ export const HeaderExample = {
         {
           type: GuideSectionTypes.JS,
           code: headerElasticPatternSource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: headerElasticPatternHtml,
         },
       ],
       text: (