+++ src-docs/src/views/page/page_example.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import { renderToHtml } from '../../services';
-
 import { GuideSectionTypes } from '../../components';
 import { pageTemplateConfig } from './playground';
 import { PageDemo } from './_page_demo';
@@ -26,61 +24,44 @@ import PageNew from './page_new';
 const pageNewSource = require('!!raw-loader!./page_new');
 import PageTemplate from './page_template';
 const PageTemplateSource = require('!!raw-loader!./page_template');
-const PageTemplateHtml = renderToHtml(PageTemplate);
 
 import PageRestricingWidth from './page_restricting_width';
 const PageRestricingWidthSource = require('!!raw-loader!./page_restricting_width');
 import PageRestricingWidthTemplate from './page_restricting_width_template';
 const PageRestricingWidthTemplateSource = require('!!raw-loader!./page_restricting_width_template');
-const PageRestricingWidthTemplateHtml = renderToHtml(
-  PageRestricingWidthTemplate
-);
 
 import PageCenteredBody from './page_centered_body';
 const PageCenteredBodySource = require('!!raw-loader!./page_centered_body');
 import PageCenteredBodyTemplate from './page_centered_body_template';
 const PageCenteredBodyTemplateSource = require('!!raw-loader!./page_centered_body_template');
-const PageCenteredBodyTemplateHtml = renderToHtml(PageCenteredBodyTemplate);
 
 import PageCenteredContent from './page_centered_content';
 const PageCenteredContentSource = require('!!raw-loader!./page_centered_content');
 import PageCenteredContentTemplate from './page_centered_content_template';
 const PageCenteredContentTemplateSource = require('!!raw-loader!./page_centered_content_template');
-const PageCenteredContentTemplateHtml = renderToHtml(
-  PageCenteredContentTemplate
-);
 
 import PageSimple from './page_simple';
 const PageSimpleSource = require('!!raw-loader!./page_simple');
 import PageSimpleTemplate from './page_simple_template';
 const PageSimpleTemplateSource = require('!!raw-loader!./page_simple_template');
-const PageSimpleTemplateHtml = renderToHtml(PageSimpleTemplate);
 
 import PageSimpleCenteredBody from './page_simple_content_body';
 const PageSimpleCenteredBodySource = require('!!raw-loader!./page_simple_content_body');
 import PageSimpleCenteredBodyTemplate from './page_simple_content_body_template';
 const PageSimpleCenteredBodyTemplateSource = require('!!raw-loader!./page_simple_content_body_template');
-const PageSimpleCenteredBodyTemplateHtml = renderToHtml(
-  PageSimpleCenteredBodyTemplate
-);
 
 import PageSimpleEmptyContent from './page_simple_empty_content';
 const PageSimpleEmptyContentSource = require('!!raw-loader!./page_simple_empty_content');
 import PageSimpleEmptyContentTemplate from './page_simple_empty_content_template';
 const PageSimpleEmptyContentTemplateSource = require('!!raw-loader!./page_simple_empty_content_template');
-const PageSimpleEmptyContentTemplateHtml = renderToHtml(
-  PageSimpleEmptyContentTemplate
-);
 
 import PageCustomContent from './page_custom_content';
 const PageCustomContentSource = require('!!raw-loader!./page_custom_content');
 import PageCustomContentTemplate from './page_custom_content_template';
 const PageCustomContentTemplateSource = require('!!raw-loader!./page_custom_content_template');
-const PageCustomContentTemplateHtml = renderToHtml(PageCustomContentTemplate);
 
 import PageLegacy from './page';
 const PageLegacySource = require('!!raw-loader!./page');
-const PageLegacyHtml = renderToHtml(PageLegacy);
 
 export const PageExample = {
   title: 'Page',
@@ -119,10 +100,6 @@ export const PageExample = {
           type: GuideSectionTypes.JS,
           code: pageNewSource,
           displayName: 'Components JS',
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageTemplateHtml,
         },
       ],
       text: (
@@ -219,10 +196,6 @@ export const PageExample = {
           code: PageRestricingWidthSource,
           displayName: 'Components JS',
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageRestricingWidthTemplateHtml,
-        },
       ],
       text: (
         <>
@@ -279,10 +252,6 @@ export const PageExample = {
           type: GuideSectionTypes.JS,
           code: PageCenteredBodySource,
           displayName: 'Components JS',
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageCenteredBodyTemplateHtml,
         },
       ],
       text: (
@@ -341,10 +310,6 @@ export const PageExample = {
           code: PageCenteredContentSource,
           displayName: 'Components JS',
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageCenteredContentTemplateHtml,
-        },
       ],
       text: (
         <>
@@ -398,10 +363,6 @@ export const PageExample = {
           code: PageSimpleSource,
           displayName: 'Components JS',
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageSimpleTemplateHtml,
-        },
       ],
       text: (
         <>
@@ -448,10 +409,6 @@ export const PageExample = {
           code: PageSimpleCenteredBodySource,
           displayName: 'Components JS',
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageSimpleCenteredBodyTemplateHtml,
-        },
       ],
       text: (
         <p>
@@ -495,10 +452,6 @@ export const PageExample = {
           code: PageSimpleEmptyContentSource,
           displayName: 'Components JS',
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageSimpleEmptyContentTemplateHtml,
-        },
       ],
       text: (
         <p>
@@ -539,10 +492,6 @@ export const PageExample = {
           code: PageCustomContentSource,
           displayName: 'Components JS',
         },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageCustomContentTemplateHtml,
-        },
       ],
       text: (
         <>
@@ -581,10 +530,6 @@ export const PageExample = {
         {
           type: GuideSectionTypes.JS,
           code: PageLegacySource,
-        },
-        {
-          type: GuideSectionTypes.HTML,
-          code: PageLegacyHtml,
         },
       ],
       text: (